requirements.txt
Django==4.2.14
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1